Refactor rendering available views
Currently, the trace-explorer-views-widget shares the same code
as the trace-overview-selection-component that renders the list
of available outputs. This commit extracts the common code into
a separate component that can be shared by both components.
